refactor the folder location
